hardening: 严格告警清零 (-Wshadow/-Wconversion/...) + 网关残留连接修复
make strict: -Wshadow -Wformat=2 -Wundef -Wconversion
-Wsign-conversion -Wwrite-strings -Wpointer-arith -Wcast-align 零警告

修复 22 处:
- audio: snprintf 返回值/RIFF 头字段显式转换 + WAV 大小上限
- zip: ctx 改 const void*; zlib uInt 转换 guard
- rbin/md: 索引显式 (size_t)
- mcp: endpoint_path 长度检查 + memcpy (防截断)
- gateway: 转发 write 返回值处理
- 测试: 字面量/ssize_t 转换

关键 bug: 网关转发时客户端提前断开 → 响应体未读完的
连接带 chunked 残留入池 → 下一请求读到 '0\r\n\r\n'
status=0 转发给客户端 (严格模式 -O1 下稳定复现)。
修复: write 失败 → pool_discard 作废连接。

保留 env 开关调试 (HTTP_DEBUG/GW_DEBUG) 供线上排查。

+++ src/data/rbin.c
@@ -52,7 +52,7 @@ static uint32_t rd_u32(Rd *r) {
 static uint64_t rd_u64(Rd *r) {
     if (r->off + 8 > r->len) { r->err = 1; return 0; }
     uint64_t v = 0;
-    for (int i = 0; i < 8; i++) v |= ((uint64_t)r->p[r->off + i]) << (8 * i);
+    for (int i = 0; i < 8; i++) v |= ((uint64_t)r->p[r->off + (size_t)i]) << (8 * i);
     r->off += 8;
     return v;
 }
